requirements.txt
allure-pytest==2.13.5
anyio==3.6.2
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.3
databases==0.7.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.109.1
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.2
orjson==3.9.15
psycopg2-binary==2.9.6
pydantic==1.10.7
pytest==8.2.2
pytest-html==4.1.1
pytest-xdist==3.6.1
python-dotenv==1.0.0
python-multipart==0.0.7
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.35.1
typing==3.7.4.3
typing_extensions==4.12.2
ujson==5.7.0
urllib3==1.26.18
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
